Refactor create sub-modules relay and p2p
