+++ mithril-relay/src/lib.rs
@@ -1,6 +1,10 @@
-pub mod client;
-pub mod peer;
-pub mod relay_aggregator;
-pub mod relay_signer;
+pub mod p2p;
+pub mod relay;
+
+pub use p2p::peer;
+pub use relay::client;
+pub use relay::AggregatorRelay;
+pub use relay::PassiveRelay;
+pub use relay::SignerRelay;
 
 pub const MITHRIL_SIGNATURES_TOPIC_NAME: &str = "mithril/signatures";
