+++ mithril-relay/src/relay/mod.rs
@@ -1,0 +1,7 @@
+pub mod aggregator;
+pub mod passive;
+pub mod signer;
+
+pub use aggregator::AggregatorRelay;
+pub use passive::PassiveRelay;
+pub use signer::SignerRelay;
